combined support classes for rosestacker and wildstacker
