+++ build.gradle.kts
@@ -17,7 +17,6 @@ repositories {
     maven("https://hub.jeff-media.com/nexus/repository/jeff-media-public/")
     maven("https://jitpack.io")
     maven("https://repo.rosewooddev.io/repository/public/")
-    maven("https://repo.rosewooddev.io/repository/public/")
     maven("https://repo.bg-software.com/repository/api/")
     mavenCentral()
 }
@@ -36,6 +35,7 @@ dependencies {
     compileOnly("me.clip:placeholderapi:2.11.5")
     compileOnly("dev.rosewood:rosestacker:1.5.17")
     compileOnly("com.bgsoftware:WildStackerAPI:2023.3")
+    compileOnly("com.github.OfficialRell:SpawnerMeta:22.5")
 
     implementation("de.jeff_media:SpigotUpdateChecker:1.3.2")
     implementation("org.bstats:bstats-bukkit:3.0.2")
